patch: move EquipmentConfig struct to it's own module
